requirements.txt
fastapi[standard]==0.121.1; python_version >= '3.8'
uvicorn[standard]==0.38.0; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'

pyjwt==2.10.1; python_version >= '3.9'
bcrypt==5.0.0; python_version >= '3.8'

sqlalchemy==2.0.44; python_version >= '3.7'
psycopg2==2.9.11; python_version >= '3.9'
alembic==1.17.1; python_version >= '3.10'

elasticsearch[async]==8.19.2; python_version >= '3.8'
elastic-transport==8.17.1; python_version >= '3.8'

pydantic[email]==2.12.4; python_version >= '3.9'
python-dotenv==1.2.1; python_version >= '3.9'
sqlmodel==0.0.27